fixed problems in the service class

+++ src/main/java/com/oburnett127/jobsearch/model/CreateRequest.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class CreateRequest {
     private String title;
-    private long employerId;
+    private int employerId;
     private String desc;
 }
